requirements.txt
jsonfield
djangorestframework
django-rest-swagger